Created the shelter details page.
Each shelter thumbnail links to the Details component. The Details then splits up incoming data between animals and shelters with the searchType parameter found in the url (react router). The api requests are made, then the images are output in the Details component. The text outputs are different, so they are made in AnimalDetails and ShelterDetails.