Lifts the state from Form components to Search page
This commit has the Search page send a function as a prop to the form
components. This function changes the state in the Search page instead
the components. This allows the state to be saved when switching between
forms.

This commit also reorganizes the logic used for the selector classnames.