Adds search bar to header
This commit adds a search bar that utilizes a built in trie. This trie
matches animal breeds stored on the site, allowing for lightning fast
lookup speeds.

This commit also revamps the Router setup. Now, all page variables are
derived from the url, not passed through the `this.props.location.state`.
This allows users to return to a page with the same information as before.